pyproject.toml
[project]
name = "domain-chatbot"
version = "0.1.0"
description = "First Agentic AI project "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "google-genai>=1.63.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
